[UPDATE] Corrections to the fast 3D intersection algorithm
The fast 3D intersection algorithm has been debugged, and tests
indicate that the algorithm is linear in complexity. The algorithm is a
slower than the 2D version  - which should be expected, given the
higher complexity of the 3D intersection operation. The latter does
~4e5 intersection operations in ~70s, while the former does the same
number of operations in ~160s.

Both measurements were done in a MacBook Air, though, so the algorithm
should be considerably faster in most realistic usages. For an example
of the timings, see the “times_2D.txt” and “times_3D.txt” files.