[ADD] 2D mesh parser
 — The 2D mesh parser was finalised and corrected, and the
corresponding code was commented. Two functions were implemented:
ImportGmsh() and ExportGmsh(), which respectively read and write a mesh
in Gmsh’s format (.msh files)

 — The “intersection_2D” program was altered to import two Gmsh files
and apply Gander’s algorithm to them. The usage is

      intersection_2D [first file] [second file]

If the files are not given, the program will use the files
“data/test_mesh_A.msh” and “data/test_mesh_B.msh” are used.

 — IMPORTANT NOTE: for now, we only need the geometrical information
from the mesh (nodes and basic triangular elements), and so any tags
and other information (for example, extra nodes used in higher order
triangles and other types of elements) is ignored. As a result, the
exported meshes will not contain this info, even if they were generated
from a file containing them.