[ADD] Intersection 2D / gmsh output
 — The intersection algorithm can now save the intersections as a
triangulation, compatible with Gmsh. Please note that, while CGAL’s
triangulation data structure was used to do this, the triangulation
built is NOT valid (no neighbour relations between the triangles are
set,
 and no infinite faces are created), but since we do not need to do
intersection operations on it, this should not be a problem (for now …)

 — The program “intersection_2D” was modified to export the
intersections. Now, it accepts a third argument, giving the filename of
the output. If no name is given, the output is set to
"data/test_mesh_intersection.msh”. The usage is now

      intersection_2D [first input file] [second input file] [output
file]