[ADD] 3D mesh intersection / 3D Arlequin formulation / lumping method
Several changes are applied with this commit:

 --- Implemented a 3D mesh intersection method using CGAL’s Nef
polyhedrons
(http://doc.cgal.org/latest/Nef_3/classCGAL_1_1Nef__polyhedron__3.html)
instead of the triangle intersections, since they can be unreliable
when using inexact construction kernels.

     Since exact constructions are needed, it is simpler to use Nef
polyhedrons for the intersections (the intersection operation itself is
a one line operation) These polyhedrons can also be used to represent
any convex polyhedron, and thus using them should make it easier to
generalize the algorithms to hexahedron elements.

 --- Implemented the Arlequin algorithm using FreeFem++. The FreeFem++
scripts and example meshes are inside the "FE3D" folder. The
"CArl*.edp" files contain the scripts themselves, and the
"runscript_FE3D*.sh" files are bash scrtipts to run them. Usage of this
script is identical to the 2D scripts. The boundary conditions are
associated to the variables "boundAX1", "boundAY1", "boundAZ1" ... and
so on.

 --- Implementation of "lumping" versions of the FreeFem++ scripts
(named "CArl_lumping.edp" for both 2D and 3D cases). These versions
apply a Direct Mass Lumping to the intersection coupling matrix. This
matrix must be inverted to apply the LATIN method. In general, this
operation is very expensive, but the "lumped" matrix is diagonal,
resulting in a trivial inversion. The resulting methods have a slower
convergence (considerably so for the 2D case), but this cost is largely
offset by the simplified inversion.