[FEATURE] Added a "restart" feature to the LATIN solver + minor changes
 - Added a “restart” feature to the LATIN solver. It will save the
convergence data, the (temporary) solutions sol_A, sol_B, and the
Lagrange multipliers \phi_dA, \phi_dB into binary files when given the
flag “Save_restart_data” in the config files. It will start the LATIN
iterations from these files with the “Use_restart_data” flag. The
folder where to save these files can be chosen using the option
“LATINRestartDataFiles”.

 - Changed a bit the output of the LATIN solver to print more
information about the matrices and the timing of the solvers.

 - [NOT OPTIMISED] Added a code to repartition the meshes in a way
better suited to solve the modified systems of equations created by the
LATIN method. This code is not optimised yet!