[UPDATE] Finished and checked the CG solver
 - Finished the interface of the CG solver for the coupled systems.

 - [IMPORTANT] I still need to implement the preconditioner (for now,
it is equal to 1). Since the convergence of the CG algorithm applied to
a system A*x = b depends on the condition number of the matrix A, and
since A contains the inverses of the system matrices, its convergence
without a preconditioner is horrible.

 - Added a generic wrapper for the model solvers, allowing the
construction of the interface with external solvers in the future.