[BUG] Fixed a syncing issue when using local Vecs
 - Corrected a syncing issue that appeared when calling PETSc's VecMAXPY().
The issue was due to rounding errors when perfoming the (same) sequential
matrix and vector operations. Now, the corresponding code does the calculations
on the first processor and broadcast the result. It should not be a performance
bottleneck since the matrices and vectors involved have rather small dimensions
(6 x 6 dense and 6, respectivelly).